+++ lib/routes/exports.dart
@@ -1,10 +1,19 @@
-export 'package:cloth_shopping_app/views/screens/auth/forgot_password.dart';
-export 'package:cloth_shopping_app/views/screens/auth/login_page.dart';
-export 'package:cloth_shopping_app/views/screens/auth/new_password.dart';
-export 'package:cloth_shopping_app/views/screens/auth/otp_page.dart';
-export 'package:cloth_shopping_app/views/screens/auth/pass_reset_success.dart';
-export 'package:cloth_shopping_app/views/screens/auth/profile_setup.dart';
-export 'package:cloth_shopping_app/views/screens/auth/signup_page.dart';
+export 'package:cloth_shopping_app/constants/colors.dart';
+export 'package:cloth_shopping_app/views/components/app_styles.dart';
+export 'package:cloth_shopping_app/views/components/custom_otp_page.dart';
+export 'package:cloth_shopping_app/views/components/elevated_button.dart';
+export 'package:cloth_shopping_app/views/components/text_form_field.dart';
+export 'package:cloth_shopping_app/views/screens/auth/login/login_page.dart';
+export 'package:cloth_shopping_app/views/screens/auth/resetpassword/forgot_password.dart';
+export 'package:cloth_shopping_app/views/screens/auth/resetpassword/new_password.dart';
+export 'package:cloth_shopping_app/views/screens/auth/resetpassword/pass_reset_success.dart';
+export 'package:cloth_shopping_app/views/screens/auth/resetpassword/reset_otp_page.dart';
+export 'package:cloth_shopping_app/views/screens/auth/signup/profile_setup.dart';
+export 'package:cloth_shopping_app/views/screens/auth/signup/signup_page.dart';
+export 'package:cloth_shopping_app/views/screens/auth/signup/verify_email.otp.dart';
 export 'package:cloth_shopping_app/views/screens/notifications/notification_permission.dart';
 export 'package:cloth_shopping_app/views/screens/onboarding/onboarding.dart';
 export 'package:cloth_shopping_app/views/screens/onboarding/select_lang.dart';
+export 'package:flutter/material.dart';
+export 'package:gap/gap.dart';
+export 'package:image_picker/image_picker.dart';
