+++ lib/views/screens/auth/resetpassword/forgot_password.dart
@@ -4,7 +4,7 @@ import 'package:cloth_shopping_app/views/components/app_styles.dart';
 import 'package:cloth_shopping_app/views/components/back_button.dart';
 import 'package:cloth_shopping_app/views/components/elevated_button.dart';
 import 'package:cloth_shopping_app/views/components/text_form_field.dart';
-import 'package:cloth_shopping_app/views/screens/auth/otp_page.dart';
+import 'package:cloth_shopping_app/views/screens/auth/resetpassword/reset_otp_page.dart';
 import 'package:flutter/material.dart';
 import 'package:gap/gap.dart';
