+++ pubspec.yaml
@@ -40,6 +40,10 @@ dependencies:
   url_launcher: ^6.3.1
   provider: ^6.1.2
   email_validator: ^3.0.0
+  image_picker: ^1.1.2
+  intl: ^0.20.0
+  hive: ^2.2.3
+  shared_preferences: ^2.2.3
 
 dev_dependencies:
   flutter_test:
